Updated spr environment
First version that is technically working. Open TODOs include comparing the DQN results with the baseline approach by Schneider et al., defining the node observations, adding support for multi-step node observations to graph-marl (start with an int that indicates the number of message passing iterations), identifying an env config that is challenging for the baseline, and finally evaluating netmon on that config.